requirements (1).txt
graphviz
google-generativeai
PyPDF2
requests
pandas
matplotlib
seaborn
plotly
streamlit
beautifulsoup4 
